添加水平角的均值计算方式，并计算水平角和俯仰角的标准差

+++ src/software/SfM/tx.h
@@ -22,5 +22,5 @@ double getDistance(double tx, double ty, double tz, double rx, double ry, double
 double getFuYang(double tx, double ty, double tz, double rx, double ry, double rz);
 double getShuiPing(double tx, double ty, double tz, double rx, double ry, double rz);
 bool sortdes(const cv::line_descriptor::KeyLine &k1, const cv::line_descriptor::KeyLine &k2);
-double Count(double a[], int size, double x);
+int Count(double a[], int size, double x);
 std::vector<cv::line_descriptor::KeyLine> drawLines(const char* imageName, const char* outputImgName);
